+++ src/Aggregator.DataAccess/Entities/Abstract/ICurrencyAmount.cs
@@ -8,7 +8,7 @@ public interface ICurrencyAmount
     /// <summary>
     /// Сумма в минимальных единицах валюты
     /// </summary>
-    public long? Amount { get; set; }
+    public decimal? Amount { get; set; }
     
     /// <summary>
     /// Трехзначный числовой код валюты (ISO-4217)
